+++ pymatgen/io/vasp/tests/test_outputs.py
@@ -4,8 +4,6 @@
 
 from __future__ import division, unicode_literals
 
-from monty.json import MontyDecoder
-
 """
 Created on Jul 16, 2012
 """
@@ -30,11 +28,9 @@ import xml.etree.cElementTree as ET
 from pymatgen.electronic_structure.core import OrbitalType
 from pymatgen.io.vasp.inputs import Kpoints
 from pymatgen.io.vasp.outputs import Chgcar, Locpot, Oszicar, Outcar, \
-    Vasprun, Procar, Xdatcar, Dynmat, BSVasprun, UnconvergedVASPWarning, NMRChemicalShiftNotation
+    Vasprun, Procar, Xdatcar, Dynmat, BSVasprun, UnconvergedVASPWarning
 from pymatgen import Spin, Orbital, Lattice, Structure
 from pymatgen.entries.compatibility import MaterialsProjectCompatibility
-import copy
-
 
 test_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..", "..",
                         'test_files')
@@ -639,81 +635,6 @@ class DynmatTest(unittest.TestCase):
             d.data[4][2]['dynmat'][3], [0.055046, -0.298080, 0.]
         ))
         # TODO: test get_phonon_frequencies once cross-checked
-
-
-class TestChemicalShiftNotation(unittest.TestCase):
-    def test_nmr_chemical_shift_notation(self):
-        cs1 = NMRChemicalShiftNotation.from_maryland_notation(
-            195.0788, 68.1733, 0.8337)
-        hae1 = cs1.haeberlen_values
-        self.assertAlmostEqual(hae1.sigma_iso, 195.0788, places=5)
-        self.assertAlmostEqual(hae1.delta_sigma, 65.33899505250002, places=5)
-        self.assertAlmostEqual(hae1.zeta, 43.559330035000016, places=5)
-        self.assertAlmostEqual(hae1.eta, 0.13013537835511454, places=5)
-        meh1 = cs1.mehring_values
-        self.assertAlmostEqual(meh1.sigma_iso, 195.0788, places=5)
-        self.assertAlmostEqual(meh1.sigma_11, 170.46483003499998, places=5)
-        self.assertAlmostEqual(meh1.sigma_22, 176.13343993, places=5)
-        self.assertAlmostEqual(meh1.sigma_33, 238.63813003500002, places=5)
-        mary1 = cs1.maryland_values
-        self.assertAlmostEqual(mary1.sigma_iso, 195.0788, places=5)
-        self.assertAlmostEqual(mary1.omega, 68.1733, places=5)
-        self.assertAlmostEqual(mary1.kappa, 0.8337, places=5)
-        cs2 = copy.deepcopy(cs1)
-        cs2.sigma_22 = 235.0
-        hae2 = cs2.haeberlen_values
-        self.assertAlmostEqual(hae2.sigma_iso, 214.70098669, places=5)
-        self.assertAlmostEqual(hae2.delta_sigma, -66.35423498250003, places=5)
-        self.assertAlmostEqual(hae2.zeta, -44.23615665500003, places=5)
-        self.assertAlmostEqual(hae2.eta, 0.08224335724674214, places=5)
-        cs2.sigma_11 = 200.0
-        cs2.sigma_33 = 270.0
-        hae3 = cs2.haeberlen_values
-        self.assertAlmostEqual(hae3.sigma_iso, 235.0, places=5)
-        self.assertAlmostEqual(hae3.delta_sigma, -52.5, places=5)
-        self.assertAlmostEqual(hae3.zeta, -35.0, places=5)
-        self.assertAlmostEqual(hae3.eta, 1.0, places=5)
-        cs2.sigma_11 = 235.0
-        hae4 = cs2.haeberlen_values
-        self.assertAlmostEqual(hae4.sigma_iso, 246.66666666666666, places=5)
-        self.assertAlmostEqual(hae4.delta_sigma, 35.0, places=5)
-        self.assertAlmostEqual(hae4.zeta, 23.333333333333343, places=5)
-        self.assertAlmostEqual(hae4.eta, 0.0, places=5)
-        cs3 = copy.deepcopy(cs1)
-        cs3.sigma_11 = 1.0
-        cs3.sigma_22 = 1000.0
-        cs3.sigma_33 = 1000.0
-        mary2 = cs3.maryland_values
-        self.assertAlmostEqual(mary2.sigma_iso, 667.0, places=5)
-        self.assertAlmostEqual(mary2.omega, 999.0, places=5)
-        self.assertAlmostEqual(mary2.kappa, -1.0, places=5)
-        cs3.sigma_22 = 1.0
-        mary3 = cs3.maryland_values
-        self.assertAlmostEqual(mary3.sigma_iso, 334.0, places=5)
-        self.assertAlmostEqual(mary3.omega, 999.0, places=5)
-        self.assertAlmostEqual(mary3.kappa, 1.0, places=5)
-        cs3.sigma_22 = 500.5
-        mary4 = cs3.maryland_values
-        self.assertAlmostEqual(mary4.sigma_iso, 500.5, places=5)
-        self.assertAlmostEqual(mary4.omega, 999.0, places=5)
-        self.assertAlmostEqual(mary4.kappa, 0.0, places=5)
-        cs4 = NMRChemicalShiftNotation(68.1733, 0.8337, 195.0788)
-        cs5 = NMRChemicalShiftNotation(0.8337, 195.0788, 68.1733)
-        self.assertAlmostEqual(cs5.sigma_11, cs4.sigma_11)
-        self.assertAlmostEqual(cs5.sigma_22, cs4.sigma_22)
-        self.assertAlmostEqual(cs5.sigma_33, cs4.sigma_33)
-        cs6 = NMRChemicalShiftNotation(195.0788, 68.1733, 0.8337)
-        self.assertAlmostEqual(cs6.sigma_11, cs4.sigma_11)
-        self.assertAlmostEqual(cs6.sigma_22, cs4.sigma_22)
-        self.assertAlmostEqual(cs6.sigma_33, cs4.sigma_33)
-        d1 = cs1.as_dict()
-        cs7 = NMRChemicalShiftNotation.from_dict(d1)
-        hae7 = cs7.haeberlen_values
-        self.assertAlmostEquals(hae1, hae7, places=5)
-        j1 = cs1.to_json()
-        cs8 = json.loads(j1, cls=MontyDecoder)
-        hae8 = cs8.haeberlen_values
-        self.assertAlmostEquals(hae1, hae8, places=5)
 
 
 if __name__ == "__main__":
